fix: 메시지 컨버터 Config 수정
