--- baguni-batch/src/main/java/baguni/batch/config/XmlMapperConfig.java
@@ -1,18 +1,0 @@
-package baguni.batch.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.dataformat.xml.XmlMapper;
-
-@Configuration
-public class XmlMapperConfig {
-
-	@Bean
-	public XmlMapper xmlMapper() {
-		XmlMapper xmlMapper = new XmlMapper();
-		xmlMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false); // 명시되지 않은 속성 무시
-		return xmlMapper;
-	}
-}
